net: shell: Add ping command
Move "ping" command to a separate file.
