+++ subsys/net/lib/shell/CMakeLists.txt
@@ -18,6 +18,7 @@ zephyr_library_sources(ipv4.c)
 zephyr_library_sources(ipv6.c)
 zephyr_library_sources(mem.c)
 zephyr_library_sources(nbr.c)
+zephyr_library_sources(ping.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
